Make iterateObserversSafely a shared utility function.
