+++ src/core/ObservableQuery.ts
@@ -10,6 +10,7 @@ import {
   Observer,
   ObservableSubscription
 } from '../utilities/observables/Observable';
+import { iterateObserversSafely } from '../utilities/observables/observables';
 import { ApolloError } from '../errors/ApolloError';
 import { QueryManager } from './QueryManager';
 import { ApolloQueryResult, FetchType, OperationVariables } from './types';
@@ -628,17 +629,4 @@ export class ObservableQuery<
 
 function defaultSubscriptionObserverErrorCallback(error: ApolloError) {
   invariant.error('Unhandled error', error.message, error.stack);
-}
-
-function iterateObserversSafely<E, A>(
-  observers: Set<Observer<E>>,
-  method: keyof Observer<E>,
-  argument?: A,
-) {
-  // In case observers is modified during iteration, we need to commit to the
-  // original elements, which also provides an opportunity to filter them down
-  // to just the observers with the given method.
-  const observersWithMethod: Observer<E>[] = [];
-  observers.forEach(obs => obs[method] && observersWithMethod.push(obs));
-  observersWithMethod.forEach(obs => (obs as any)[method](argument));
 }
